refactor: clean results view
Made-with: Cursor
